feat: corrige delete

+++ src/django_project/video_app/repository.py
@@ -38,7 +38,8 @@ class DjangoORMVideoRepository(VideoRepository):
                 video_model.categories.set(video.categories)
                 video_model.genres.set(video.genres)
                 video_model.cast_members.set(video.cast_members)
-                video_model.video.delete()
+                if video_model.video:
+                    video_model.video.delete()
 
                 video_model.video = VideoMediaModel.objects.create(
                     name=video.video.name,
